Eliminate the `useBackupFolders` hook
