+++ lib/plottr_components/src/components/dashboard/backups/BackupsHome.js
@@ -5,22 +5,20 @@ import { Col, Grid, Row, FormControl } from 'react-bootstrap'
 import UnconnectedDashboardErrorBoundary from '../../containers/DashboardErrorBoundary'
 import UnconnectedBackupFiles from './BackupFiles'
 import UnconnectedFolders from './Folders'
-import { checkDependencies } from '../../checkDependencies'
 
 const BackupsHomeConnector = (connector) => {
-  const {
-    platform: { useBackupFolders },
-  } = connector
-  checkDependencies({ useBackupFolders })
-
   const BackupFiles = UnconnectedBackupFiles(connector)
   const Folders = UnconnectedFolders(connector)
   const DashboardErrorBoundary = UnconnectedDashboardErrorBoundary(connector)
 
-  const BackupsHome = ({ userId, settings }) => {
+  const BackupsHome = ({ userId, settings, computeFolders }) => {
     const [selectedFolder, selectFolder] = useState(null)
     const [searchTerm, setSearchTerm] = useState('')
-    const folders = useBackupFolders(userId, searchTerm, !selectedFolder)
+    const [folders, setFolders] = useState([])
+
+    useEffect(() => {
+      setFolders(computeFolders(searchTerm, selectedFolder))
+    }, [searchTerm, selectedFolder, setFolders, computeFolders])
 
     useEffect(() => {
       setSearchTerm('')
@@ -80,6 +78,7 @@ const BackupsHomeConnector = (connector) => {
   BackupsHome.propTypes = {
     userId: PropTypes.string.isRequired,
     settings: PropTypes.object.isRequired,
+    computeFolders: PropTypes.func.isRequired,
   }
 
   const {
@@ -93,6 +92,8 @@ const BackupsHomeConnector = (connector) => {
     return connect((state) => ({
       userId: selectors.userIdSelector(state.present),
       settings: selectors.appSettingsSelector(state.present),
+      computeFolders: (searchTerm, selectedFolder) =>
+        selectors.filteredSortedBackupsSelector(state.present, searchTerm, selectedFolder),
     }))(BackupsHome)
   }
