+++ lib/plottr_components/src/connections/pltr.js
@@ -190,7 +190,6 @@ const pltrTypeSpecs = {
     export_config: PropTypes.object.isRequired,
     saveExportConfigSettings: PropTypes.func.isRequired,
   }),
-  useBackupFolders: PropTypes.func.isRequired,
   moveFromTemp: PropTypes.func.isRequired,
   showItemInFolder: PropTypes.func.isRequired,
   tempFilesPath: PropTypes.string.isRequired,
